Block Matching Stereo: maxError in native formats
- Fixed maxError being type-cast to an int breaking it in many scenarios
- Examples: large squared error, images normalized to have a pixel value of 1, ...

+++ main/boofcv-ip-multiview/src/main/java/boofcv/alg/disparity/block/select/SelectErrorSubpixel.java
@@ -37,7 +37,7 @@ public class SelectErrorSubpixel {
 		/// If the error is a squared error. If false then it's assumed to be distance
 		final boolean squaredError;
 
-		public S32_F32( int maxError, int rightToLeftTolerance, double texture, boolean squaredError ) {
+		public S32_F32( double maxError, int rightToLeftTolerance, double texture, boolean squaredError ) {
 			super(maxError, rightToLeftTolerance, texture, GrayF32.class);
 			this.squaredError = squaredError;
 		}
@@ -86,7 +86,7 @@ public class SelectErrorSubpixel {
 		/// If the error is a squared error. If false then it's assumed to be distance
 		final boolean squaredError;
 
-		public F32_F32( int maxError, int rightToLeftTolerance, double texture, boolean squaredError ) {
+		public F32_F32( double maxError, int rightToLeftTolerance, double texture, boolean squaredError ) {
 			super(maxError, rightToLeftTolerance, texture, GrayF32.class);
 			this.squaredError = squaredError;
 		}
